requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
pydantic-settings
python-multipart
jinja2
docling
transformers
torch
accelerate
pillow
onnxruntime